moved styles in seperate files
