+++ src/navigation/navigation.tsx
@@ -1,5 +1,3 @@
-// Navigation.tsx
-
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
