+++ src/screens/PrivacyPolicyScreen/index.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { View, Text, TouchableOpacity, StyleSheet, SafeAreaView, Image, ScrollView } from 'react-native';
-import { fontFamily } from '../constant/fontFamily';
-import { ImagePaths } from '../constant/images';
+import { View, Text, TouchableOpacity, SafeAreaView, Image, ScrollView } from 'react-native';
+import { ImagePaths } from '../../constant/images';
 import { useNavigation } from '@react-navigation/native';
-import { RootStackParamList } from '../interfaces/common';
-
+import { RootStackParamList } from '../../interfaces/common';
+import { styles } from './styles';
 
 const PrivacyPolicyScreen: React.FC= () => {
   const navigation = useNavigation<RootStackParamList>();
@@ -36,42 +35,6 @@ const PrivacyPolicyScreen: React.FC= () => {
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor:'#FFFFFF'
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding:20,
-    backgroundColor:'#F2F2F2'
-  },
-  headerText: {
-    fontSize: 15,
-    fontFamily: fontFamily.bold,
-    color:'#0D0D0D'
-  },
-  closeButton: {
-    padding: 10,
-    position:'absolute',
-    right:20
-  },
-  closeIcon: {
-    width:15,
-    height:15
-  },
-  content: {
-    padding:30,
-    paddingBottom:60
-  },
-  paragraph: {
-    fontSize: 13,
-    lineHeight: 21,
-    color:'#2A2A2A',
-    marginBottom:30
-  },
-});
+
 
 export default PrivacyPolicyScreen;
